fix: polling for events in startHealthy never ends
`startHealthy` wasn't giving an `until` filter to docker, so the stream
polling for event was never ending, causing client processes not to terminate.

New algorithm does filter on `until`, and will only poll for up to one second
beyond the point where the timeout elapses or the host healthy signal is
received.